Reduced redundancies and updates LSP settings
The trend for LSP plugins is for them to now have one for each
langserver available that then downloads the dependencies needed for
it to run. This means we don't need to manually install _nearly_
as much stuff. Hurray. Also included various small changes that have
accumulated over time.

Note that Exalt is now disabled not because I dislike the package, but
rather because it's not supported on M1 Macs and that's what I'm
actively using at this time.